web-panel/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
jinja2
python-multipart
httpx
python-dotenv
starlette
